その他指摘箇所の修正
